requirements.txt
factory_boy
Django==2.1
django-dotenv==1.4.2
django-model-utils==3.1.2
mysqlclient==1.3.13
pytz==2018.5
selenium==3.14.0
urllib3==1.23
